Update runtime to 23.08, do some housekeeping and add comments for clarity.
I would have split this into multiple commits, but it was easier to just dump the changes all at once.

When built this works on my system (Fedora Rawhide with AMD Graphics) for hardware acceleration, and there aren't any visible issues with the frontend.

I did some changes to how the flatpak is built, as it's better to not ship AR with the final flatpak. The final flatpak only ships the needed libraries and the binary.